[Debt] Fixed PluginData registration of Bridge
Created dagger modules for each platform and turned Bridge abstract.
Injections to classes that require Plugin specific objects are not
possible, and will lead to compile time errors.

Classes that require information about another class at runtime cause
Plan to fail to start if the class is not present - another solution has
to be worked for PlaceholderAPI.