ConfigNode copy now preserves the original properly
Fixed a possible future issue where tree up-traversal is broken
if a node in a tree is passed to a copy operation.
This was because the parent node was updated in the process of copy if
no old node was present.

Fixed by creating a new node when copying instead of reusing the
object being copied