Patch against bad AFKThreshold setting data
Patch that resets AFK time of sessions with afk time of length of
the session to 0.

This is a bug remedy patch that attempts to turn "bad" afk data to good.
In 4.5.2 there was a bug that caused some config setting defaults not
being copied, along those AFKThreshold setting,
which lead to AFK threshold being read as 0.
This in turn lead to full sessions being regarded as having been AFK.