Reworked WorldAliasSettings to not use a Map
It does not make any sense to copy the values to a Map when they are
already stored in memory in a tree like structure accessible in similar
manner.

Prevents NPE in getAliases() if World_aliases contains two level setting
1:
  14: "1.14"

Now the alias is read properly even with dots in the name.

Accidental fix to Most played world String:
"Unknown" makes no sense for a world alias if the actual name of the
world is known. Thus it is now displayed instead.

Affects issues:
- Close #1089