Removed Kingdoms support
I'm guessing the plugin moved to Songonda at some point
and Songonda is now off Spigot, so a free version is not
available to use the API from.

Since I am not able to obtain a copy anymore, the bridge
can not be built - I need to build it to fix a bug, so
the support was removed.