Removed Most/Least Active time (Too ambiguous)
A single stat saying "Most Active Time is Sunday, 2 AM" unfortunately
does not convey what I originally would have liked in the context it is
presented in.

This is because it is not possible to easily figure out what day
session occurred - The sessions are right next to the stat so this would
bring confusion as the user is not able to easily verify the result.

This stat might be re-introduced to Online Activity Overview tab some
time in the future, since the day names are easily checked from
Punchcard or Calendar.