Added an ISO-8601 Timezone independent formatter
Server calendar had an issue with days going on
the wrong date if timezone was negative GMT.

Fixed the issue by implementing a timezone
independent ISO-8601 formatter to use with the dates on
the session calendar. The date given to ServerCalendar is already
set to midnight on each day.

Affects issues:
- Fixed #1239