All Extension tables now displayed
Original query used an odd inner join which caused some tables to
not be displayed at all, in some cases no tables were displayed at all.
The inner join was joining the value table, which is lower in the table hierarchy.

Changed the query to instead inner join upwards in the table hierarchy, joining the
plugin table.

Affects issues:
- #1260