Rewrote the way server information is loaded on enable
- The UUID of the server (and other info) is stored in a file
  on Proxy servers as well to speed up loading time.
- The information is stored in two places, with preference for the info in
  the file. The information is updated after the plugin has enabled.

This opens a possibility for a bug, where user changes UUID inside
Bungee file, leading to the Bungee storing a second BungeeCord entry to
the database. This can lead to issues. If this comes up the file
can be changed to different format (to deter users from modifying it manually)

Affects issues:
- Close #1247
  - Fixed #1478
  - Fixed #1496
  - Fixed #1637