Fixed offset for graphs with minutes in timezones
Timezone offset is now a double.

This fix seems very simple, and it kinda is.

HighCharts represents offset in minutes.
Java gives offset in milliseconds, and that was
incorrectly to be assumed as full hours, so the offset
was rounded to an integer. Using a double will allow
all kinds of offsets that still work with highcharts.

Fixes timezones like
- Nepal (GMT+05:45)
- Newfoundland Time (GMT-03:30)

Affects issues:
- Fixed #1652