Hacky fix to forge crash
The server crashes due to sqlite driver loading disabling class loading caches.
This re-enables them after using the method.

Note that if a class is loaded in a small timeframe between sqlite driver loading and
the finally-block running there is a slight chance of a crash, but it is less than 100%

Affects issues:
- Fix #2202