Revert most changes to existing html from i18n PR merge
The changes break some things that are difficult to fix and effort is better spent rewriting in React.

Most important changes are already incorporated into the React rewrite,
but these changes prevent releasing the completed parts of the rewrite in a dev release under a feature flag.

Note that the paths of the different locale lines (Lang) still need to be looked through before merge.