Add plan.command permission, default: true
This permission is not enforced and only
affects Bukkit based servers.

The permission allows negating visibility
of the command in tab completion for players

The subcommands rely on other permissions,
so 'true' was set as default and permission
unenforced so that previous functionality
of the commands stays as is.

Affects issues:
- Fix #2441