Fix issue with sidebar closing right away on first click
Sidebar items not being rendered caused the currentTab to be undefined,
until the sidebar is opened, and then currentTab is set - which then
caused a useEffect that closes sidebar to fire.
This lead to the sidebar closing right away on the first click.

To fix this the sidebar item state was moved to navigationHook,
and changing the list of items on a sidebar updates currentTab
at the same time to prevent sidebar rendering setting it again.