Override default timezone in TZIndependentFormatter as UTC
The formatter was applying Operating System timezone on top of any offset,
leading to incorrect dates for any values passed to it.

This meant that anything at the start of day at negative GMT timezones, eg. GMT-4
would move back one day, showing incorrect results in the calendar.

Fixed by applying UTC as the timezone for the formatter

Affects issues:
- Possibly fixed #1239